Models directory was renamed to ViewModels
