+++ VacationRental.Api.Tests/GetCalendarTests.cs
@@ -2,7 +2,7 @@
 using System.Net;
 using System.Net.Http;
 using System.Threading.Tasks;
-using VacationRental.Api.Models;
+using VacationRental.Api.ViewModels;
 using Xunit;
 
 namespace VacationRental.Api.Tests
