+++ VacationRental.Api/Controllers/RentalsController.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
 using Microsoft.AspNetCore.Mvc;
-using VacationRental.Api.Models;
+using VacationRental.Api.ViewModels;
 
 namespace VacationRental.Api.Controllers
 {
