--- VacationRental.Api/Models/BookingViewModel.cs
@@ -1,12 +1,0 @@
-﻿using System;
-
-namespace VacationRental.Api.Models
-{
-    public class BookingViewModel
-    {
-        public int Id { get; set; }
-        public int RentalId { get; set; }
-        public DateTime Start { get; set; }
-        public int Nights { get; set; }
-    }
-}
