+++ VacationRental.Api/Startup.cs
@@ -5,7 +5,7 @@ using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using Microsoft.Extensions.Hosting;
 using Microsoft.OpenApi.Models;
-using VacationRental.Api.Models;
+using VacationRental.Api.ViewModels;
 
 namespace VacationRental.Api
 {
