+++ VacationRental.Api/ViewModels/BookingBindingModel.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 
-namespace VacationRental.Api.Models
+namespace VacationRental.Api.ViewModels
 {
     public class BookingBindingModel
     {
