+++ VacationRental.Api/ViewModels/CalendarViewModel.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace VacationRental.Api.Models
+namespace VacationRental.Api.ViewModels
 {
     public class CalendarViewModel
     {
