+++ VacationRental.Api/ViewModels/RentalBindingModel.cs
@@ -1,4 +1,4 @@
-﻿namespace VacationRental.Api.Models
+﻿namespace VacationRental.Api.ViewModels
 {
     public class RentalBindingModel
     {
