+++ VacationRental.Api/ViewModels/RentalViewModel.cs
@@ -1,4 +1,4 @@
-﻿namespace VacationRental.Api.Models
+﻿namespace VacationRental.Api.ViewModels
 {
     public class RentalViewModel
     {
